services/backend/requirements.txt
asyncpg==0.25.0
databases[postgresql]==0.5.3
fastapi==0.70.0
psycopg2-binary==2.9.2
SQLAlchemy==1.4.27
email-validator==1.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
uvicorn==0.15.0
python-decouple==3.5
python-multipart==0.0.5
# dev
pytest==6.2.5
requests==2.26.0
